+++ src/components/index.ts
@@ -13,3 +13,4 @@ export { default as Settings } from './Settings/Settings';
 export { default as Share } from './Share/Share';
 export { default as Status } from './Status/Status';
 export { default as Tracker } from './Tracker/Tracker';
+export { default as UpdateSW } from './UpdateSW/UpdateSW';